+++ arrays/src/main/kotlin/IntroArrays.kt
@@ -1,21 +1,5 @@
+import teste.testaAlteracaoElementosArray
+
 fun main() {
-    val idades: IntArray = intArrayOf(25, 19, 33, 20, 55, 40)
-
-    var maiorIdade = Int.MIN_VALUE
-    for (idade in idades) {
-        if (idade > maiorIdade) {
-            maiorIdade = idade
-        }
-    }
-
-    println(maiorIdade)
-
-    var menorIdade = Int.MAX_VALUE
-    idades.forEach { idade ->
-        if (idade < menorIdade) {
-            menorIdade = idade
-        }
-    }
-
-    println(menorIdade)
+    testaAlteracaoElementosArray()
 }
